+++ bocadillo/__init__.py
@@ -10,4 +10,4 @@ from .websockets import WebSocket, WebSocketDisconnect
 from .response import Response
 from .request import Request
 
-__version__ = "0.10.1"
+__version__ = "0.10.2"