+++ bocadillo/misc.py
@@ -7,3 +7,11 @@ assets_dir = join(dirname(abspath(__file__)), "assets")
 def read_asset(filename: str) -> str:
     with open(join(assets_dir, filename), "r") as f:
         return f.read()
+
+
+def overrides(original):
+    def decorate(func):
+        func.__doc__ = original.__doc__
+        return func
+
+    return decorate