+++ bocadillo/websockets.py
@@ -1,4 +1,3 @@
-import json
 from typing import Awaitable, Callable, Optional, Any, Union, Tuple
 
 from starlette.datastructures import URL
@@ -7,45 +6,8 @@ from starlette.websockets import (
     WebSocketDisconnect as _WebSocketDisconnect,
 )
 
-from .app_types import Event
+from .app_types import Event, Scope, Receive, Send
 from .constants import WEBSOCKET_CLOSE_CODES
-
-_STARLETTE_WEBSOCKET_DOCS = (
-    "[Starlette.websockets.WebSocket](https://www.starlette.io/websockets/)"
-)
-
-
-def _get_alias_docs(name: str) -> str:
-    return f"\n\nAlias of `{name}` on {_STARLETTE_WEBSOCKET_DOCS}."
-
-
-class _Delegated:
-    """Descriptor to delegate a method to the underlying Starlette WebSocket.
-
-    # See Also
-    - [Descriptors](https://docs.python.org/3/reference/datamodel.html#implementing-descriptors)
-    """
-
-    def __init__(self, websocket_attr: str = "_websocket"):
-        self.websocket_attr = websocket_attr
-        self._docs_imported = False
-
-    def __set_name__(self, owner, name: str):
-        # Use the declared attribute's name as source attribute name.
-        self.source = name
-
-    def __get__(self, instance: Optional["WebSocket"], owner):
-        if instance is None:  # pragma: no cover
-            # Class attribute access.
-            # NOTE: used by Pydoc-Markdown when generating docs.
-            obj = getattr(StarletteWebSocket, self.source)
-            if not self._docs_imported:
-                obj.__doc__ = (obj.__doc__ or "") + _get_alias_docs(self.source)
-                self._docs_imported = True
-            return obj
-        else:
-            # Instance attribute access.
-            return getattr(getattr(instance, self.websocket_attr), self.source)
 
 
 class WebSocket:
@@ -75,7 +37,9 @@ class WebSocket:
 
     def __init__(
         self,
-        *args,
+        scope: Scope,
+        receive: Receive,
+        send: Send,
         value_type: Optional[str] = None,
         receive_type: Optional[str] = None,
         send_type: Optional[str] = None,
@@ -86,7 +50,7 @@ class WebSocket:
         # WebSocket class uses those in many other functions, which we
         # do not need / want to re-implement.
         # The compromise is the definition of delegated methods below.
-        self._websocket = StarletteWebSocket(*args)
+        self._ws = StarletteWebSocket(scope, receive=receive, send=send)
 
         if caught_close_codes is None:
             caught_close_codes = (1000, 1001)
@@ -99,51 +63,97 @@ class WebSocket:
         else:
             receive_type = receive_type or self.__default_receive_type__
             send_type = send_type or self.__default_send_type__
+
         self.receive_type = receive_type
         self.send_type = send_type
 
-    # Methods delegated to the underlying Starlette WebSocket object.
-    # TODO: add type annotations.
-    accept = _Delegated()
-    close = _Delegated()
-    receive_text = _Delegated()
-    send_text = _Delegated()
-    receive_bytes = _Delegated()
-    send_bytes = _Delegated()
-
     @property
     def url(self) -> URL:
-        return self._websocket.url
+        """The URL which the WebSocket connection was made to.
 
-    async def receive_json(self) -> Union[dict, list]:
-        """Return `json.loads(await self.receive_text())`.
-
-        # Returns
-        json (dict or list): a JSON message.
+        # See Also
+        - [starlette.WebSocket.url](https://www.starlette.io/websockets/#url)
         """
-        # Starlette decodes from bytes by default,
-        # but most WebSocket clients generally send text.
-        return json.loads(await self.receive_text())
+        return self._ws.url
 
-    async def send_json(self, message: Union[dict, list]):
-        """Execute `await self.send_text(json.dumps(message))`.
+    # Connection handling.
+
+    async def accept(self, subprotocol: str = None) -> None:
+        """Accept the connection request.
 
         # Parameters
-        message (list or dict): a JSON message.
+        subprotocol (str): a specific WebSocket subprotocol that should be expected from clients. Defaults to `None`.
         """
-        # Encodes as bytes by default,
-        # but most WebSocket clients don't expect to receive plain bytes.
-        await self.send_text(json.dumps(message))
+        return await self._ws.accept(subprotocol=subprotocol)
+
+    async def reject(self):
+        """Reject the connection request.
+
+        This is equivalent to `.close(403)`.
+        Calling this before `.accept()` has undefined behavior.
+        """
+        return await self.close(code=403)
+
+    async def close(self, code: int = 1000) -> None:
+        """Close the WebSocket connection.
+
+        # Parameters
+        code (int): a close code, defaults to `1000` (Normal Closure).
+
+        # See Also
+        - [CloseEvent specification](https://developer.mozilla.org/en-US/docs/Web/API/CloseEvent)
+        """
+        return await self._ws.close(code=code)
+
+    async def ensure_closed(self, code: int = 1000):
+        """Close the connection if it has not been closed already."""
+        try:
+            await self.close(code=code)
+        except RuntimeError:
+            # Already closed.
+            pass
+
+    # Data reception/sending.
+
+    async def receive_text(self) -> str:
+        """Receive a message as text."""
+        return await self._ws.receive_text()
+
+    async def send_text(self, data: str):
+        """Send a message as text."""
+        return await self._ws.send_text(data)
+
+    async def receive_bytes(self) -> bytes:
+        """Receive a message a bytes."""
+        return await self._ws.receive_bytes()
+
+    async def send_bytes(self, data: bytes):
+        """Send a message as bytes."""
+        return await self._ws.send_bytes(data)
+
+    async def receive_json(self) -> Union[dict, list]:
+        """Receive a message as text and parse it to a JSON object.
+
+        # Raises
+        json.JSONDecodeError: if the received JSON is invalid.
+        """
+        return await self._ws.receive_json()
+
+    async def send_json(self, data: Union[dict, list]):
+        """Serialize an object to JSON and send it as text.
+
+        # Raises
+        TypeError: if the given `data` is not JSON serializable.
+        """
+        return await self._ws.send_json(data)
 
     async def receive_event(self) -> Event:
-        return await self._websocket.receive()
-
-    receive_event.__doc__ = _get_alias_docs("receive")
+        """Receive a raw ASGI event."""
+        return await self._ws.receive()
 
     async def send_event(self, event: Event):
-        return await self._websocket.send(event)
-
-    send_event.__doc__ = _get_alias_docs("send")
+        """Send a raw ASGI event."""
+        return await self._ws.send(event)
 
     async def receive(self) -> Union[str, bytes, list, dict]:
         """Receive a message from the WebSocket.
@@ -161,26 +171,6 @@ class WebSocket:
         sender = getattr(self, f"send_{self.send_type}")
         return await sender(message)
 
-    async def ensure_closed(self, code: int = 1000):
-        """Close the connection if it has not been closed already.
-
-        # Parameters
-        code (int): a close code, defaults to `1000`.
-        """
-        try:
-            await self.close(code)
-        except RuntimeError:
-            # Already closed.
-            pass
-
-    async def reject(self):
-        """Reject the connection request.
-
-        This is equivalent to `await close(403)`.
-        Calling this before `accept()` has undefined behavior.
-        """
-        return await self.close(code=403)
-
     # Asynchronous context manager.
 
     async def __aenter__(self, *args, **kwargs):
@@ -193,12 +183,12 @@ class WebSocket:
             # Returning `True` here silences the exception. See:
             # https://docs.python.org/3/reference/datamodel.html#object.__exit__
             return exc_val.code in self.caught_close_codes
-        else:
-            # Close with Internal Error if an exception was raised.
-            code = 1000 if exc_type is None else 1011
-            # NOTE: view may have closed the WebSocket already, so we
-            # must use ensure_closed().
-            await self.ensure_closed(code)
+
+        # Close with Internal Error if an unknown exception was raised.
+        code = 1000 if exc_type is None else 1011
+        # NOTE: view may have closed the WebSocket already, so we
+        # must use ensure_closed().
+        await self.ensure_closed(code)
 
     # Asynchronous iterator.